p8fs-api/pyproject.toml
[project]
name = "p8fs-api"
version = "1.2.6"
description = "REST API, CLI, and MCP interfaces for the P8FS smart content management system"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi>=0.104.0",
    "hypercorn>=0.16.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "PyJWT[crypto]>=2.10.1",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "httpx>=0.25.0",
    "structlog>=23.2.0",
    "loguru>=0.7.2",
    "prometheus-client>=0.19.0",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-instrumentation-fastapi>=0.42b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.21.0",
    "opentelemetry-instrumentation-httpx>=0.42b0",
    "opentelemetry-instrumentation-logging>=0.42b0",
    "slowapi>=0.1.9",
    "fastmcp>=2.13.0",
    "starlette>=0.49.1",
    "p8fs-cluster",
    "p8fs",
    "p8fs-auth",
    "qrcode[pil]>=8.2",
    "jinja2>=3.1.6",
    "boto3>=1.34.0",
    "rich>=14.2.0",
    "authlib>=1.6.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.25.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.uv.sources]
p8fs-cluster = { workspace = true }
p8fs = { workspace = true }
p8fs-auth = { workspace = true }

[dependency-groups]
dev = [
    "asgi-lifespan>=2.1.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.1.0",
]
